// src/app/woven/models/sample.ts
export class Sample {
    samplename: any;
    companyid: any;
    brandid: any;
    programid: any;
    imagepath: any;
    productcategoryid: any;
    productsubcategoryid: any;
    clientid: any;
    executiveid: any;
    designerid: any;
    salesid: any;
    orderdate: any;
    estdeliverydate: any;
    length: any;
    width: any;
    pcsyard: any;
    genspcStatus: any;
    expiryStatus: boolean;
    isbarcode: boolean;
    issubcontact: number;
    isjobformultimaterial: boolean;
    approveldate: any;
    groundcolor: any;
    textcolor: any;
    submissiondate: any;
    submissiondate2: any;
    submissiondate3: any;
    submissiondate4: any;
    submissioncomment: any;
    submissioncomment2: any;
    submissioncomment3: any;
    submissioncomment4: any;
    rejectiondate: any;
    rejectioncomment: any;
    rejectiondate2: any;
    rejectiondate3: any;
    rejectiondate4: any;
    rejectioncomment2: any;
    rejectioncomment3: any;
    rejectioncomment4: any;
    remark1: any;
    specialinstruction: any;
    artworklocation: any;
    unit: any;
    filename: any;
    partno: any;
    papergsm: any;
    expireDate: any;
    deliverydate: any;
    createdby: any;
    completeDate: any;
    finishlength: any;
    reasonOfChange: any;
    totalnoofcolor: any;

    constructor() {
        this.samplename = '';
        this.companyid = 0 ;
        this.brandid = 0;
        this.programid = 0;
        this.clientid = 0;
        this.imagepath = '';
        this.executiveid = 0;
        this.finishlength = 0;
        this.designerid = 0;
        this.papergsm = 0;
        this.salesid = 0;
        this.unit = 0;
        this.orderdate = '';
        this.estdeliverydate = '';
        this.specialinstruction = '';
        this.groundcolor = '';
        this.textcolor = '';
        this.pcsyard = '' ;
        this.genspcStatus = '';
        this.expiryStatus = false;
        this.isbarcode = false;
        this.isjobformultimaterial = false;
        this.issubcontact = 0;
        this.approveldate = '';
        this.expireDate = '';
        this.completeDate = '';
        this.deliverydate = '';
        this.artworklocation = '';
        this.filename = '';
        this.submissiondate =  '';
        this.submissiondate2 = '';
        this.submissiondate3 =  '';
        this.submissiondate4 =  '';
        this.submissioncomment =  '';
        this.submissioncomment2 =  '';
        this.submissioncomment3 =  '';
        this.submissioncomment4 =  '';
        this.rejectioncomment =  '';
        this.partno = 1;
        this.createdby = 0;
        this.rejectiondate = '';
        this.rejectioncomment2 =  '';
        this.rejectioncomment3 =  '';
        this.rejectioncomment4 =  '';
        this.rejectiondate2 =  '';
        this.rejectiondate3 = '';
        this.rejectiondate4 = '';
        this.reasonOfChange = '';
        this.totalnoofcolor = 0;
    }
}

export class SampleWiseKeyEntryField {
  sampleid: any;
  keyentry1: any;
  keyentry2: any;
  keyentry3: any;
  keyentry4: any;
  keyentry5: any;
  keyentry6: any;
  keyentry7: any;
  keyentry8: any;
  keyentry9: any;
  keyentry10: any;
  statusid: any;
  createdby: any;
    creationdate: any;
  constructor() {
    this.sampleid = 0;
  }
}